se_praktikumsgruppe9_backend/src/requirements.txt
pytest

numpy>=1.24.2
scipy>=1.10.1
pandas>=1.5.3
django==4.1.7

djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2